Refactor HudiUtil.getSplits to HudiSplitFactory
